[refactor] ESLint 수정
